+++ typescript/vrotsc/e2e/cases/all-content/policyTemplates/PolicyTemplateSchedule.pl.ts
@@ -3,10 +3,19 @@ import { PolicyTemplate } from "vrotsc-annotations";
 @PolicyTemplate({
 	name: "Policy Template Schedule",
 	path: "MyOrg/MyProject",
-	schedule: {
-		periode: "every-days",
-		when: "11:12:00,12:00:00",
-		timezone: "Europe/Sofia"
+	templateVersion: "v2",
+	elements: {
+		Schedule: {
+			type: "Periodic Event",
+			events: {
+				OnExecute: "onExecute"
+			},
+			schedule: {
+				periode: "every-days",
+				when: "11:12:00,12:00:00",
+				timezone: "Europe/Sofia"
+			}
+		}
 	}
 })
 export class PolicyTemplateAmqp {